# ext/bcmath: Made the same changes to `_bc_do_add` as `_bc_do_sub`