@SakiTakamachi 2024-05-12
@nielsdos 
I noticed this with Travis.
https://app.travis-ci.com/github/php/php-src/builds/270410835

The other CIs certainly did not detect this.

This is due to a rear overrun, but I can't think of any test to detect this... Do you have any good ideas?