@SakiTakamachi 2024-05-12
Hmm, I'll merge this for now and add tests later if possible.